+++ Apptentive/Apptentive/Engagement/Model/ApptentiveDevice.h
@@ -91,11 +91,6 @@ extern NSString *const ATDeviceLastUpdateValuePreferenceKey;
 @property (copy, nonatomic) NSDictionary *integrationConfiguration;
 
 /**
- The advertising identifier, if the AdSupport framework is linked, and the user has enabled it
- */
-@property (readonly, nullable, strong, nonatomic) NSUUID *advertisingIdentifier;
-
-/**
  Initializes a device object with values obtained from the current device.
 
  @return The newly-initialized device object.
@@ -113,11 +108,6 @@ extern NSString *const ATDeviceLastUpdateValuePreferenceKey;
 + (void)getPermanentDeviceValues;
 
 /**
- Sets static variable for advertising identifier.
- */
-+ (void)getAdvertisingIdentifier;
-
-/**
  The push integration to be set globally for all devices
  */
 @property (class, strong, nonatomic) NSDictionary *integrationConfiguration;
@@ -133,11 +123,6 @@ extern NSString *const ATDeviceLastUpdateValuePreferenceKey;
  This is an intermediate value that device objects can use to update themselves.
  */
 @property (class, strong, nonatomic) UIContentSizeCategory contentSizeCategory;
-
-/**
- Exposes the static variable _currentAdvertisingIdentifier for reading when debugging.
- */
-@property (class, readonly, strong, nonatomic) NSUUID *advertisingIdentifier;
 
 @end